Increasing retries and delay to make 500 errors causes by Datastore eventual consistency less likely to happen.
Details:
When table entity doesn’t exist, then we create proper table entity, and based on that entity we schedule proper copy job.

Based on that copy job couldn’t be scheduled before creation of table entity, after copy job assumes that table entity exist.

Not having table entity is treaded as an error → if we really don’t have table entity for backup what we made, then it means that we have inconsistent data in Datastore.

In existing cases, from eventual consistency of Datastore we don’t get existing table_entity and we retried request till the success.

Existing errors shown that this request was retried 5 times before task fails.
Fortunately, whole task was retried and 3rd attempt of retry was successful.

As error is not repeatitable (occurs randomly in time, probably being dependent from datastore servers load), I increased number of retries and make it in longest period of time to make this kind of errors less likely to happen.