YACHT-964: Refactor is_daily_partitioned method
Previously it performs two checks:
- if table has daily partitioning set in metadata
- if table is not single partition (as single partition has also set daily partitioning)

It could produce bugs, as in some cases we could want to know if table is daily partitioned but we don't care if table is single partition or main table. In other cases we want to know if the table is single partition.

We decided to divide it to different methods.